opencl_mandelbrot: cl: Replace all double constants with floats
When offline compiling clang would generate Double or Float64 types
which make the ICD complain.

+++ opencl_mandelbrot/mandelbrot.cl
@@ -22,7 +22,7 @@ cfloat mandelbrot_step(cfloat z, cfloat c)
 
 inline uint mod2color(float mod, float mod_max)
 {
-    const float scale = (255.0 / mod_max) * mod;
+    const float scale = (255.0f / mod_max) * mod;
     return 0x010001 * ((uint)scale);
 }
 
@@ -31,7 +31,7 @@ __kernel void mandelbrot(
     const uint height,
     __global uint* bitmap
 ) {
-    const ulong bitmap_index_end = width * height;
+    const uint bitmap_index_end = width * height;
 
     const int local_id = get_global_id(0);
     if (local_id > bitmap_index_end)
@@ -42,25 +42,25 @@ __kernel void mandelbrot(
 
     cfloat pos = (cfloat)(local_col, local_row);
 
-    const float zoom = 1.15;
+    const float zoom = 1.15f;
 
     /* Scale X */
     pos.x /= width;
-    pos.x = pos.x*3.0 - 2.5;
+    pos.x = pos.x*3.0f - 2.5f;
     pos.x *= zoom;
 
     /* Scale Y */
     pos.y /= height;
-    pos.y = pos.y*2.0 - 1.0;
+    pos.y = pos.y*2.0f - 1.0f;
     pos.y *= zoom;
 
-    cfloat fout = (cfloat)(0.0, 0.0);
+    cfloat fout = (cfloat)(0.0f, 0.0f);
     for (uint i = 0; i < 32; ++i) {
         fout = mandelbrot_step(fout, pos);
     }
 
-    const float cutoff = 0.85;
-    const float fout_mod = clamp(cmod(fout), (float)0.0, cutoff);
+    const float cutoff = 0.85f;
+    const float fout_mod = clamp(cmod(fout), 0.0f, cutoff);
 
     const uint color = mod2color(fout_mod, cutoff);
     bitmap[local_id] = color;
